# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.AuthDetails:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
    type: object
  auth.RefreshInput:
    properties:
      fingerprint:
        type: string
      refreshToken:
        type: string
    type: object
...
  domain.UserInfo:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
...
  dto.CreateUserDTO:
    properties:
      cartID:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  dto.SignInDTO:
    properties:
      email:
        type: string
      fingerprint:
        type: string
      password:
        type: string
    type: object
  dto.SignUpDTO:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  dto.UpdateUserDTO:
    properties:
      cartID:
        type: string
      name:
        type: string
    type: object
  v1.failure:
    properties:
      error:
        $ref: '#/definitions/v1.failureInfo'
    type: object
  v1.failureInfo:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: invalid request body
        type: string
    type: object
  v1.success:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact:
    email: paw1a@yandex.ru
    name: API Support
    url: https://t.me/paw1a
  description: This is simple api of e-commerce shop
  title: E-commerce API
  version: "1.0"
paths:
  /admins/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: refresh services
        in: body
        name: refreshInput
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthDetails'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Admin refresh token
      tags:
      - admin-auth
  /admins/auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: admin credentials
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/dto.SignInDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthDetails'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: Admin sign-in
      tags:
      - admin-auth
...
  /admins/users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.success'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get all users
      tags:
      - admin-users
    post:
      consumes:
      - application/json
      parameters:
      - description: user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Create user
      tags:
      - admin-users
  /admins/users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Delete user
      tags:
      - admin-users
    get:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Get user by id
      tags:
      - admin-users
    put:
      consumes:
      - application/json
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: string
      - description: user update fields
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - AdminAuth: []
      summary: Update user
      tags:
      - admin-users
...
  /users/account:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthDetails'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - UserAuth: []
      summary: User account
      tags:
      - user
  /users/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: user refresh services
        in: body
        name: refreshInput
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthDetails'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: User refresh token
      tags:
      - user-auth
  /users/auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: user credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.SignInDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthDetails'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: User sign-in
      tags:
      - user-auth
  /users/auth/sign-up:
    post:
      consumes:
      - application/json
      parameters:
      - description: user services
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.failure'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      summary: User sign-up
      tags:
      - user-auth
...
  /users/reviews:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v1.success'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.failure'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.failure'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.failure'
      security:
      - UserAuth: []
      summary: User reviews List
      tags:
      - user
securityDefinitions:
  AdminAuth:
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"